pyproject.toml
[project]
name = "dumbmoney"
version = "0.2.1"
description = "Technical analysis and quantitative trading toolkit for retail investors taking on the market"
readme = "README.md"
authors = [
  { name = "littledumb", email = "littledumb@dumbsmart.org" }
]
license = "MIT"
requires-python = ">=3.9"
dependencies = [
    "akshare>=1.17.91",
    "massive>=2.0.2",
    "mplfinance>=0.12.10b0",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "tushare>=1.4.24",
]

[project.urls]
Homepage = "https://github.com/dumbsmart-org/dumbmoney"
Repository = "https://github.com/dumbsmart-org/dumbmoney.git"
Issues = "https://github.com/dumbsmart-org/dumbmoney/issues"

[build-system]
requires = ["uv_build>=0.9.13,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "httpx[socks]>=0.28.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "pytest-mpl>=0.18.0",
    "python-dotenv>=1.2.1",
]

[[tool.uv.index]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
publish-url = "https://test.pypi.org/legacy/"
explicit = true
